refactor(ipc): first slice of #696 — shared constants, struct dedup, EventEmitter trait
Move constants (INFERENCE_FAILURE_MESSAGES, IDLE_MESSAGES,
NPC_REACTION_CONCURRENCY, REQUEST_ID with SeqCst) from per-backend
files into parish-core::ipc::handlers so all three runtimes share a
single source of truth.

Move shared state structs (ConversationRuntimeState, SaveState,
UiConfigSnapshot) to parish-core::ipc::state; update parish-server and
parish-tauri to re-export from there instead of duplicating.

Re-export IPC payload types (StreamTokenPayload, StreamTurnEndPayload,
StreamEndPayload, TextLogPayload, NpcReactionPayload, LoadingPayload)
from parish-core in parish-tauri/src/events.rs, replacing local copies.

Add object-safe EventEmitter trait stub to parish-core::ipc::event_emitter
(send+sync, single emit_event method); not yet wired to callers — that
is the next slice.

Fix 13 TextLogPayload struct literals in parish-tauri/src/commands.rs
that were missing the subtype field after the unified struct definition
was adopted.

Partial fix for #696. Full refactor is multi-PR.
