+++ parish/crates/parish-core/src/ipc/mod.rs
@@ -7,7 +7,9 @@
 pub mod commands;
 pub mod config;
 pub mod editor;
+pub mod event_emitter;
 pub mod handlers;
+pub mod state;
 pub mod streaming;
 pub mod types;
 
@@ -15,6 +17,8 @@ pub use commands::{
     CommandEffect, CommandResult, TextPresentation, handle_command, render_look_text,
 };
 pub use config::GameConfig;
+pub use event_emitter::EventEmitter;
 pub use handlers::*;
+pub use state::{ConversationRuntimeState, SaveState, UiConfigSnapshot};
 pub use streaming::{TOKEN_CHANNEL_CAPACITY, stream_npc_tokens};
 pub use types::*;
